avancement cpp09 - fix cpp06
uintptr = c99 et le c99 - même si fonctionnelle sur compilateur récent - n'est pas dans la norme c++98 . L'énoncé de l'exo est faux !

+++ CPP/CPP06/ex01/Serializer.hpp
@@ -23,8 +23,8 @@ private:
 
 public:
 	/////////////////////////////////////////////////////////////////////////// méthodes
-	static uintptr_t serialize(Data* ptr);
-	static Data* deserialize(uintptr_t raw);
+	static unsigned long serialize(Data* ptr);
+	static Data* deserialize(unsigned long raw);
 
 };
